requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.4.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
